Issues with updateTask
Got a better data structure onto Firebase, added a new file in order to help read in that data as well

I'm 100% positive that theres something wrong with the data that i'm putting into the Task class. Because every time it goes to updateDisplayedTask it kills the app. We're going to need that so it can refresh the page and add it in

also modified variables in Task.java